test: split the guard spec by concern (#85)
Break `safePolicyGuard.spec.ts` into one file per concern and lift its
fixture into a shared module.

## Context and Motivation

At 1,315 lines and 42 tests across eleven top-level blocks, the file
covered everything from interface IDs to delayed configuration to
execution rollback. Its size is what made the earlier `enableGuard`
migration a PR of its own, and it is the last item in the test-suite
refactor.

## Changes

- `test/fixtures.ts`: the fixture, as `safePolicyGuardFixture`.
- `safePolicyGuard.spec.ts` (244 lines, 10 tests): the constructor, the
interface surface, the two guard entry points, and guard removal --
which drives the delayed path end to end and is kept whole rather than
split across two files.
- `safePolicyGuardConfiguration.spec.ts` (350, 10):
`configureImmediately`.
- `safePolicyGuardDelayedConfiguration.spec.ts` (568, 15):
`requestConfiguration`, `applyConfiguration`, `invalidateRoot`.
- `safePolicyGuardContext.spec.ts` (84, 3): context decoding.
- `safePolicyGuardExecution.spec.ts` (93, 4): execution outcome.

## Decisions and Tradeoffs

- Configuration is split in two rather than kept as one lifecycle file,
which would have been 887 lines and close to the problem being solved.
The seam is a real distinction in the contract: `configureImmediately`
is the pre-guard bootstrap path, while request/apply/invalidate is the
delayed path available once the guard is installed, and the delay is the
mechanism itself.
- The fixture is imported as `safePolicyGuardFixture as fixture`, so
every `loadFixture(fixture)` call site is untouched and the split stays
reviewable as a move rather than a rewrite.
- `ReentrantMockPolicyMode` goes only to the execution spec, the sole
consumer.

## Testing

`npm test` 120 passing, unchanged; lint and typecheck green. Because a
mechanical split can silently drop a test, this was checked three ways
rather than on the total alone: the 42 test names extracted from the
previous revision diff identically against the new files with no
renames; the per-file counts sum to 42 (10 + 10 + 15 + 3 + 4); and a
line-by-line comparison of test content shows the only lines absent from
the new files are the inline fixture that moved. The blocks were
extracted programmatically rather than by hand to avoid transcription
drift.
